refactor: reorganize keys logic into modular files
