+++ handlers/keys/__init__.py
@@ -4,6 +4,10 @@ from aiogram import Router
 
 from .key_management import router as management_router
 from .keys import router as keys_router
+from .key_view import router as view_router
+from .key_renew import router as renew_router
+from .key_freeze import router as freeze_router
+from .key_connect import router as connect_router
 
 
 router = Router(name="keys_main_router")
@@ -11,4 +15,8 @@ router = Router(name="keys_main_router")
 router.include_routers(
     keys_router,
     management_router,
+    view_router,
+    renew_router,
+    freeze_router,
+    connect_router
 )
